DecorationsWidgets: use double for scale property
(bzr r3680.1.73)

+++ decorations/DecorationsWidgets.h
@@ -50,7 +50,7 @@ public:
   nux::Property<bool> focused;
   nux::Property<bool> sensitive;
   nux::Property<bool> mouse_owner;
-  nux::Property<float> scale;
+  nux::Property<double> scale;
 
   CompRect const& Geometry() const;
   virtual int GetNaturalWidth() const;
